% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{convert_glyph_to_df}
\alias{convert_glyph_to_df}
\title{Convert a Hershey glyph code to a data.frame}
\usage{
convert_glyph_to_df(glyph)
}
\arguments{
\item{glyph}{glyph code character string e.g. "    8  9MWOMOV RUMUV ROQUQ"}
}
\value{
data.frame with x, y and stroke group.
}
\description{
A not particularly elegant or fast way to convert a Hershey font code string
to a data.frame.
}
