% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{convert_hershey_font_to_df}
\alias{convert_hershey_font_to_df}
\title{Read a .jhf Hershey font file as a data.frame}
\usage{
convert_hershey_font_to_df(jhf_filename)
}
\arguments{
\item{jhf_filename}{Hershey font file (.jhf)}
}
\value{
data.frame
}
\description{
Read a .jhf Hershey font file as a data.frame
}
