% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{create_string_df}
\alias{create_string_df}
\title{Create a data.frame of glyph strokes for the given string}
\usage{
create_string_df(text, font = "rowmant")
}
\arguments{
\item{text}{text string}

\item{font}{font name. default 'rowmant'}
}
\value{
data.frame with coorindates for all glyphs with characters
offset appropriately. 'char_idx' is the index of the character within the
given 'text' string.
}
\description{
An error will occur if the string includes ascii which is not part of the font.
}
