% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hershey.R
\docType{data}
\name{hershey_raw}
\alias{hershey_raw}
\title{Hershey vector fonts as a list of character strings of the original glyph encodings}
\format{An object of class \code{list} of length 33.}
\usage{
hershey_raw
}
\description{
See vignette for a description of this format.
}
\keyword{datasets}
