% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hershey.R
\docType{data}
\name{hershey_svg}
\alias{hershey_svg}
\title{Hershey vector fonts as a list of character strings of SVG paths for each glyph}
\format{An object of class \code{list} of length 33.}
\usage{
hershey_svg
}
\description{
Hershey vector fonts as a list of character strings of SVG paths for each glyph
}
\keyword{datasets}
