% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{read_jhf}
\alias{read_jhf}
\title{Read a Hershey Font file in .jhf format}
\usage{
read_jhf(jhf_filename, trim_start = TRUE)
}
\arguments{
\item{jhf_filename}{filename}

\item{trim_start}{= TRUE}
}
\value{
character vector. 1 string per glyph.
}
\description{
Read a Hershey Font file in .jhf format
}
